changed to svg mask implementation
I wanted to be able to manipulate the SVG through CSS, and it turns out
we can use a mask to do this. This masks the whole div to conform to the
SVG's shape.

mask-size: contain;
mask-position: center;
mask-repeat: no-repeat;
mask-image: url('<the url>');

With the code above, you can now change the color of the icon via the
`background-color` property in CSS.