refactor: cleaner impl of fetch/getShopifyData
I was having an issue with fetching via relative path in production, so
I figured it'll be better to just call the fn directly. It makes sense
too, why would the server have to fetch data from itself?

On the plus side, it's now cleaner code, and I don't run into issues
running the production build of this app.